Split the V110 live test out and move the two pins it broke (#2860)
Two CI failures, both pins doing their job.

#1776's hygiene rule requires every class reaching the shared
DARLING_TEST_PG store to serialize against the other live classes or
record why it does not. CollectionLogFetchPhaseSumsStoreTests was eleven
pure source-and-SQL pins plus one live migration test, so the rule's own
preferred remedy applies: the live test moves into
CollectionLogFetchPhaseSumsLivePostgresTests with
[Collection("live-postgres")] rather than serializing eleven pure pins
behind the store. Being in the collection also means the fixture has
already migrated the store, so the two MigrateAsync calls are its second
and third applications - a stronger idempotency claim than before.
ExpectedColumns and RungVersion became internal so both classes assert
against the same list rather than a second copy that could drift.

#2864's ThePeerMarkIsCapturedAtDispatchNotAtCompletion pinned
"result.Drain, peerMaxAtDispatchMs, _logger" - the argument BEFORE the
mark - so inserting result.FetchPhases there broke it. The positive half
moves. The negative half was anchored on result.Drain, which means the
insertion would have made it un-matchable and it would have gone on
passing while guarding nothing; it is re-anchored on what the defect
actually looks like, PeerMaxOrNull(server) immediately before _logger,
which is adjacency-independent and cannot be silenced by the next
insertion. A count assertion for fetchPhases: null on the seven failure
arms was added alongside the existing drain: null one.

Both proven red first, plus a failure arm attributing a fetch value
instead of NULL - twelve variants now, each on a different assertion.
36/36 green locally.
